+++ libs/qtdlib/common/qabstractint53id.h
@@ -15,7 +15,7 @@ class QAbstractInt53Id : public QAbstractTdObject
 public:
     explicit QAbstractInt53Id(QObject *parent = nullptr);
 
-    qint32 id() const;
+    qint64 id() const;
     QString qmlId() const;
 
     // QAbstractTdObject interface
